requirements.txt
# Core Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LLM & Agents
langchain==0.1.5
langchain-anthropic==0.1.6
langchain-openai==0.0.5
langgraph==0.0.20

# Document Processing
python-docx==1.1.0
PyPDF2==3.0.1
markdown==3.5.2
weasyprint==60.1

# CLI
typer==0.9.0
rich==13.7.0

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
ruff==0.2.0
mypy==1.8.0
